Delta-128 Quasar Boilerplate > Plugins > Add support for translating properties of an object (#7)
Based on this discussion: https://github.com/anyTV/freedom-partner-dashboard/pull/1#discussion_r164004740,
the boilerplate should support the translation of properties inside an object. In addition, since it already supports object, it would be nice to also support an array of keys to be translated.
However, we cannot use Vue filters for translating array or object of keys and return an array or object since filters should return a string to be able to be used for piping (e.g. 'key' | $t | uppercase ). Thus, a mixin is required to inject a method called `$trans` to accommodate our needs.

Approved by:
- @pandafox <allen.yu@freedom.tm>